pyproject.toml
[project]
name = "alpha-nex"
version = "0.1.0"
description = "Content platform with AI integration and reward systems"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "flask-wtf>=1.2.2",
]
